+++ src/java/Service/Interfaces/IAuth/IPasswordResetTokenService.java
@@ -1,7 +1,8 @@
-package Service.Interfaces;
+package Service.Interfaces.IAuth;
 
 import Exception.NotFoundException;
 import Model.Entity.OAuth.PasswordResetToken;
+import Service.Interfaces.Service;
 
 public interface IPasswordResetTokenService extends Service<PasswordResetToken> {
     public PasswordResetToken findByToken(String token) throws NotFoundException;
